tools/downloader: split off the reporting-related functionality into a module
